- Modifiche alle interfacce di Donatore e Ricevente - Aggiunti metodo in UtenteRepository.kt per prelevare un utente e relativi dati. - Aggiunti metodo in PaniereRepository.kt per prelevare i panieri a seconda della tipologia dell'utente: panieri donati per i donatori, panieri seguiti per i riceventi. - Aggiunta la View: ProfileRiceventeFragment.kt - Aggiunti i ViewModel: ProfileRiceventeFragment.kt e ProfileDonatoreViewModel.kt che fungono da Controller per le view - Aggiunta la possibilità di Switchare tra utente donatore e utente ricevente con relativa UPDATE della tipologia nel DB. - Aggiunti i vari XML relativi alle View - Rimossi pulsanti inutili - Aggiunto in fragment_profile_donatore.xml l'immagine utente, la sua reputazione visibile in termini di stelle e pulsante di Switch della tipologia. Presente inoltre una Label personalizzata che da il benvenuto all'utente mostrando il suo nome e cognome - Aggiunto nel gradle la dependency relativa a Lottie Animation - Altre piccole modifiche

+++ app/src/main/java/com/rapnap/panpar/model/Utente.kt
@@ -6,7 +6,7 @@ data class Utente (var id: String,
                    var location: String = "",
                    var tipo: Tipologia = Tipologia.DONATORE,
                    var punteggio: Long = 0, //Ricevente
-                   var rating: Long = 0, //Donatore
+                   var rating: Double = 0.0, //Donatore
                    var isNew: Boolean = true
                     )
